fix: Update function signatures to use unsigned long for hex functions and improve handling in ft_putnbr_fd and ft_putptr_fd

+++ ft_printf.h
@@ -20,8 +20,8 @@
 int		ft_putptr_fd(void *p, int fd);
 int     ft_putstr_fd(char *s, int fd);
 int 	ft_putnbr_fd(int n, int fd);
-int		ft_puthexl_fd(int n, int fd);
-int		ft_puthexu_fd(int n, int fd);
+int		ft_puthexl_fd(unsigned long n, int fd);
+int		ft_puthexu_fd(unsigned long n, int fd);
 int		ft_putchar_fd(char c, int fd);
 int		ft_printf(const char *format, ...);
